One-function-one-test-module
