+++ pallets/pallet-deposit-storage/src/deposit/tests/on_identity_committed.rs
@@ -122,122 +122,3 @@ fn on_identity_committed_insufficient_balance() {
 		);
 	});
 }
-
-#[test]
-fn on_commitment_removed_successful() {
-	let namespace = DepositNamespaces::get();
-	let key: DepositKeyOf<TestRuntime> = (SUBJECT, SUBMITTER, 0 as IdentityCommitmentVersion)
-		.encode()
-		.try_into()
-		.unwrap();
-	ExtBuilder::default()
-		.with_deposits(vec![(
-			key.clone(),
-			DepositEntry {
-				deposit: Deposit {
-					amount: 1_000,
-					owner: SUBMITTER,
-				},
-				reason: HoldReason::Deposit.into(),
-			},
-		)])
-		.build()
-		.execute_with(|| {
-			assert_eq!(
-				Pallet::<TestRuntime>::deposits(&namespace, &key),
-				Some(DepositEntry {
-					reason: HoldReason::Deposit.into(),
-					deposit: Deposit {
-						amount: 1_000,
-						owner: SUBMITTER
-					}
-				})
-			);
-			assert_eq!(
-				pallet_balances::Pallet::<TestRuntime>::balance_on_hold(&HoldReason::Deposit.into(), &SUBMITTER),
-				1_000
-			);
-
-			assert_ok!(
-				<DepositCollectorHook::<TestRuntime> as ProviderHooks<TestRuntime>>::on_commitment_removed(
-					&SUBJECT,
-					&SUBMITTER,
-					&IdentityCommitmentOf::<TestRuntime>::default(),
-					0
-				),
-			);
-
-			assert!(Pallet::<TestRuntime>::deposits(&namespace, &key).is_none(),);
-			assert!(
-				pallet_balances::Pallet::<TestRuntime>::balance_on_hold(&HoldReason::Deposit.into(), &SUBMITTER)
-					.is_zero()
-			);
-		});
-}
-
-#[test]
-fn on_commitment_different_owner_successful() {
-	let namespace = DepositNamespaces::get();
-	let key: DepositKeyOf<TestRuntime> = (SUBJECT, SUBMITTER, 0 as IdentityCommitmentVersion)
-		.encode()
-		.try_into()
-		.unwrap();
-	ExtBuilder::default()
-		.with_deposits(vec![(
-			key.clone(),
-			DepositEntry {
-				deposit: Deposit {
-					amount: 1_000,
-					owner: SUBJECT,
-				},
-				reason: HoldReason::Deposit.into(),
-			},
-		)])
-		.build()
-		.execute_with(|| {
-			assert_eq!(
-				Pallet::<TestRuntime>::deposits(&namespace, &key),
-				Some(DepositEntry {
-					reason: HoldReason::Deposit.into(),
-					deposit: Deposit {
-						amount: 1_000,
-						owner: SUBJECT
-					}
-				})
-			);
-			assert_eq!(
-				pallet_balances::Pallet::<TestRuntime>::balance_on_hold(&HoldReason::Deposit.into(), &SUBJECT),
-				1_000
-			);
-
-			assert_ok!(
-				<DepositCollectorHook::<TestRuntime> as ProviderHooks<TestRuntime>>::on_commitment_removed(
-					&SUBJECT,
-					&SUBMITTER,
-					&IdentityCommitmentOf::<TestRuntime>::default(),
-					0
-				)
-			);
-
-			assert!(Pallet::<TestRuntime>::deposits(&namespace, &key).is_none(),);
-			assert!(
-				pallet_balances::Pallet::<TestRuntime>::balance_on_hold(&HoldReason::Deposit.into(), &SUBJECT)
-					.is_zero()
-			);
-		});
-}
-
-#[test]
-fn on_commitment_removed_deposit_not_found() {
-	ExtBuilder::default().build().execute_with(|| {
-		assert_noop!(
-			<DepositCollectorHook::<TestRuntime> as ProviderHooks<TestRuntime>>::on_commitment_removed(
-				&SUBJECT,
-				&SUBMITTER,
-				&IdentityCommitmentOf::<TestRuntime>::default(),
-				0
-			),
-			FixedDepositCollectorViaDepositsPalletError::DepositNotFound
-		);
-	});
-}
